move transformerStore/ up a level
